dir_hash_c: enforce a 32 bit calculation
Yet another hashing type is just pointless - it won't increase the
randomisation.  Revert to hashing the way it used to work.

+++ lib/util.c
@@ -289,7 +289,7 @@ int dir_hash_c(const char *name, int full)
 
     if (full) {
 	unsigned char *pt;
-	unsigned long long int n;
+	uint32_t n;
 	enum {
 	    DIR_X = 3,
 	    DIR_Y = 5,
